fdo#76327: only enclose the checkbox in the sdt block.
The exporter writes the sdt blocks enclosing an entire paragraph, but
that's not the proper behaviour in all cases. The documents that mix
checkboxes and text in the same paragraph export the text inside the
sdt box and that's incorrect.

We have added code to be able to write sdt blocks that enclose
paragraphs or text runs, depending on the type of block. We have
applied it to checkboxes now.

There are two sets of properties in DocxAttributeOutput that store
sdt properties for the text run and for the paragraph. We have
modified the method WriteParagraphSdt to be generic enough to support
both cases. To write the sdt block enclosing the text run we used
parser marks.

We have renamed the property id PROP_PARA_SDTPR to the more generic
name PROP_SDTPR.

Checkbox unit test was modified to match the new structure of the
generated document.

Change-Id: I81ffe0062e1a5f80fc4638f7ee454a9bc18e63ee