ooxml: preserve rich text sdt controls
These controls don't have a special property, like for example
unformatted text controls have. So we use the id property as a
marker; we will grab-bag it together with other sdt properties and
use the existing mechanism to write the sdt block on export.

A grab bag that only contains an id property is for sure a rich text
control so we add it to the character props and not to the paragraph
props, like in the case of the unformatted text control.

Word doesn't allow us to write an empty <w:id/> tag, so we fill
it with a random number.

Finally, modified an existing unit test to add a rich text control and
check it is exported correctly.

Change-Id: If403a4a2393d4ee069a628645e364d21f104a859