Remove table-related redlines when table,row,cell removed
A table, a row or a cell might have redlines objects attached to it.
This patch makes sure than when a table\row\cell are removed -
any redlines that are attached to them are removed from the
'SwExtraRedlineTbl' object. This is to prevent any 'orphaned'
redline objects.

Conflicts:
	sw/source/core/doc/docredln.cxx
Reviewed on:
	https://gerrit.libreoffice.org/8726

Change-Id: I992e3fb4aadeb891ffd472b5d638d337a8609c01