fdo#76563 DOCX import: speed up importing lots of hyperlinks
The problem was that in
writerfilter::ooxml::OOXMLStreamImpl::lcl_getTarget(), we went over the
list of all hyperlinks for each request. Instead, let's do it once, and
in the remaining cases just look up the result from a map.

Numbers before on my machine (ms, load time):

2215, 2243, 2205

After:

1362, 1358, 1358

So that causes about 39% speedup for a non-debug build.

Change-Id: Ib4718abbe834c5ba49a85469148b656e3c808041