Bypass also the MobileLibreOffice thing for now
It seems abandoned and hasn't work since we switched to tiled
rendering anyway. And the need for the special single tile at a time
rendering test mode is mostly gone now, when tiled rendering can be
tested with the TiledLibreOffice app, I think. So no need to slow down
a "make" by building the MobileLibreOffice app, too.

Change-Id: I7b0afd3b35ff2ed0fb72f2c150abb25548a5546d