sot: fix some bizarre resize code that creates 4G temp files on 32bit
StgTmpStrm::SetSize(): the nEndOfData - 1 may underflow if nEndOfData=0
which happened to be the same value of STREAM_SEEK_TO_END before
but not any more; also it's a mystery to me why the last byte in the
buffer would need to be overwritten with 0 here.

This also fixes the failure of ww8_export test.

Change-Id: Iefe5a22c667ee83fbc315d772f357b913105ad93