sw: still use SvxBrushItem for page style background
Commit cf0185ce0eece81d6592f41f54d68e73e261903a (Merge back branch
alg_writerframes to trunk, 2014-03-19) changed the sw UNO API, so that
in case SvxBrushItem / RES_BACKGROUND was to be get/set, it mapped that
to svx's fill style.

This has two problems: 1) it does not work (see the failing unit test)
and also page style editing UI sets SvxBrushItem, so even if the UI is
used to (actually successfully) set the page background, it won't be
saved to ODF.

So, do this background <-> fill style mapping only for frames for now.

Change-Id: I734880a40c891ae9e341cbddf8f88b0462361aac