SwXTextFrame: restore semi-lost BackColor property
In the past in case FillStyle was set, BackColor wasn't ignored, now it
is, since cf0185ce0eece81d6592f41f54d68e73e261903a (Merge back branch
alg_writerframes to trunk, 2014-03-19).

The intention there was that in case FillStyle is set, FillColor will be
used: so to keep both new users and old filters happy, just restore
handling of BackColor in case FillStyle is set, but not FillColor.

Change-Id: Ie034a959e72e48b617cf598af8f7afd030bb50c4