fdo#63786 rhbz#1081968: fix PDF export of cropped SVG
When a META_FLOATTRANSPARENT_ACTION is exported, the
beginTransparencyGroup()/endTransparencyGroup() will write something
like "Q 0 0 0 rg" where the Q clears the previously set clipping region,
which is then not restored.

a68dcff60b42e06ca53f1ea6f9e675da557eedde appears to be a fix for a
previous related problem; this essentially reverts it since it
introduced the current cropping problem.  (Particularly funny is the
comment "force reemitting clip region" since the commit does the
opposite.)

It does not really make sense to me to write something to the page
content stream from beginRedirect()/endRedirect(), because that does not
actually have any effect on the _content_ of the substream, because the
referencing of the substream is only written _after_ endRedirect()
returns, so basically what is written can in the best case cancel itself
out and in the worst case introduce bugs; the state manipulations there
are just to modify the m_aCurrentPDFState following the push()/pop() (it
is necessary to clear the clip region before starting the substream to
avoid spurious "Q" at the beginning of the stream).

Change-Id: I4a5f57b784068bd7d3ba187115444574f7418195