oox: Do not overwrite table style properties
Some table properties can be defined by the table style but cells can
overwrite them in their cell properties section. Our exporter was
writing all the cell properties in all cases, regardless of them
being defined by the theme or not, and we shouldn't do that if we
want the document to work properly in Word.

To fix the issue I store the style-defined cell properties (the
format of all four borders) in the table grab bag. The exporter
recovers them and compares with the cell properties before writing
them; if the cell property matches the stlye-defined one, we don't
write it to the document.

An existing unit test was slightly modified to check that the actual
cell properties are not being skipped.

Change-Id: I3aa12d76fb8f73d3fd300f254d19e1683fb6146c