fdo#76628 RTF import: allow the same font to have multiple encodings
Commit a3aeb1463e2f8c711dd642cf7cbaa39e6a78b729 (rtftok: stop sending
sprm:CRgFtc{0,1,2} tokens, 2014-03-05) dropped support for case when a
font name is used in multiple entries in the font table, but with
different encodings.

Turns out that this is a valid use-case, so revert back to the old
behavior where the key of the encoding table is the font index, not the
font name.

Change-Id: I048dff58af801d704fd4bc75a6a4dcb0f03bf185