For consistency, set the same "GDI" and "Win" flags for iOS as for OS X
No visible effect, though. Which is both good and bad: Good because it
means that without ill effects a larger degree of idenical code paths
can be used for both iOS and OS X. Bad because this change didn't help
in getting rid of the annoying misrenderings on iOS of some Smart Art
objects.

Change-Id: I9da0f98ca90554dbac963688705b3c7955021741