Add somewhat ridiculous debug hacks for CoreGrahics drawing debugging on iOS
When the environment variable DRAW_INCREMENTALLY_FROM is set to a
number, we want TiledLibreOffice to loop, initially performing only
that number of drawing operations in AquaSalGraphics, then wait for
some seconds, and redraw. Next time perform one operation
more. Repeat.

Implemented in vcl by surrounding the entry and exit(s) of the drawing
functions in AquaSalGraphics with macros.

All this is active only for iOS and in a dbgutil build.