Refactor VCL bitmap code
Turns out that all bitmap code eventually calls on ImplDrawBitmap and
ImplDrawBitmapEx, which is called by a version with the same function
signature but without a type parameter. If you make that sal_uLong
parameter a default parameter, you can remove ImplDrawBitmap/Ex and just
call on that version of DrawBitmap/Ex.

Change-Id: Ifd75561bb824b3941c2cf700984ca80e0b851bd3