Unify IOS/OS X code in OutputDevice::DrawTransformedBitmapEx()
The way that we apply the alpha mask in DrawTransformedBitmapEx should
be handled in the same way, regardless of the platform. As it turns out,
the OS X/iOS way is quite correct and, as a bonus, the code is far more
clear. There is no need to do things differently, so let's use the OS X
approach.

Change-Id: Ib0ac72ddc549367462e84639b74e4bb148128cd5