adjust output file extension when exporting (bnc#870581)
If foo.odt is open, 'foo.odt' is the file name passed to the file dialog
for File->Export to PDF (simply because it's the document name).  If automatic
file extension is enabled, the file dialog will fix this to the PDF file type,
but if disabled, the default filename will remain foo.odt, possibly overwriting
the source file if the user is foolish enough to confirm without checking.

Change-Id: I46504630b5f4e489d5a0cedf0b2b358305763722