fdo#77759 : Embedded excel getting renamed to binary object.
The embedded excel (xlsx) when inserted on its own (unlike with chart) is getting converted to ".bin".
Note that only extension differs otherwise its an excel file only.
There is no loss of user experience in editing the file in MS Word.

The similar case is also exists when power point is inserted on it's own.

Conflicts:
	sw/qa/extras/ooxmlexport/ooxmlexport.cxx

Reviewed on:
	https://gerrit.libreoffice.org/9126

Change-Id: Ie13c098a794179c3b27100a7b9e30884fb47f656