Toolbar overflow/context menu: keep ordering and don't discard separator.
Previously the overflow/context menu should have had a separator shown
between the overflow items and the hidden items, however these were
inadvertently removed by ImplClearPopupMenu which filters the items
before the menu is displayed.

The previous ordering of items was also the REVERSE of the ordering in
the toolbar -- the overflow menu is a logical extension of the toolbar
hence items should be in the same order as they were in the toolbar.

Change-Id: I8444f4814fea64be1d8f8790445ad6aa01532e70