avoid several 200ms delays in Qt tests
Since the tests are run during LO startup, this is not such a good idea.
Also redo the socket notifier test which seems a bit of an overkill
when a mere pipe will do (and I'm not sure TCP sockets would have
the data available the moment it's written to).

Change-Id: I6a436b286d20ceecf859f9028af98da03c2561b7